+++ tests/testthat.R
@@ -61,7 +61,8 @@ if (identical(Sys.info()[["sysname"]], "Darwin")) {
   # batch 3
   c("focei-wang2007-boxcox-half", "nlm-cens", "issue-429", "issue-470",
     "focei-wang2007-bounded", "saem-loglik", "mu-timevarying", "saem-nearpd",
-    "saem-nonmutheta"),
+    "saem-nonmutheta", "saem-sharedinner", "focei-theta-reset-bounds"),
+
   # batch 4
   c("impmap", "matexp", "mfocei", "focei-wang2007-yeojohnson",
     "focei-wang2007-boxcox-lnorm", "nlme", "focei-fast-grad"),